% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropCIsGUI-package.R
\docType{package}
\name{PropCIsGUI-package}
\alias{PropCIsGUI}
\alias{PropCIsGUI-package}
\title{PropCIsGUI: Shiny Application for Confidence Intervals for Proportions}
\description{
A 'shiny' graphical interface exposing all confidence interval methods for proportions from the 'PropCIs' package. Supports single proportions (Wilson, Clopper-Pearson, mid-P, Blaker, Agresti-Coull), two independent proportions (difference, relative risk, odds ratio), matched pairs, and Bayesian credible intervals via Monte Carlo simulation. Only one method is computed at a time to prevent type I error inflation through cherry-picking of results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shearer/PropCIsGUI}
  \item Report bugs at \url{https://github.com/shearer/PropCIsGUI/issues}
}

}
\author{
\strong{Maintainer}: Ralph Scherer \email{shearer.ra76@gmail.com}

Authors:
\itemize{
  \item Ralph Scherer \email{shearer.ra76@gmail.com}
}

}
\keyword{internal}
